+++ examples/display/canvas_fonts.rb
@@ -15,29 +15,29 @@ canvas = display.canvas
 y = 10
 canvas.text_cursor = [1,y]
 canvas.font = Denko::Fonts::LED_5x7
-canvas.print("LED_5x7")
+canvas.text "LED_5x7"
 
 y += 10
 canvas.text_cursor = [0,y]
 canvas.font = Denko::Fonts::LED_6x8
-canvas.print("LED_6x8 (Default)")
+canvas.text "LED_6x8 (Default)"
 
 y += 18
 canvas.text_cursor = [1,y]
 canvas.font = Denko::Fonts::LED_8x16
-canvas.print("LED_8x16")
+canvas.text "LED_8x16"
 
 y += 20
 canvas.text_cursor = [0,y]
 canvas.font = Denko::Fonts::LED_6x8
 canvas.font_scale = 2
-canvas.print("LED_6x8")
+canvas.text "LED_6x8"
 
 canvas.font_scale = 1
 canvas.text_cursor = [86,y-8]
-canvas.print(" (2x) ")
+canvas.text " (2x) "
 canvas.text_cursor = [84,y]
-canvas.print(" scale ")
+canvas.text " scale "
 
 display.draw
 board.finish_write