+++ examples/display/canvas_transforms.rb
@@ -18,26 +18,26 @@ canvas = display.canvas
 # Very useful reflected text...
 canvas.reflect(:x)
 canvas.text_cursor = [32, 32]
-canvas.print("Relfected!")
+canvas.text "Relfected!"
 
 # Cancels out #reflex(:x) above, returning to the original coordinates.
 canvas.reflect(:x)
 canvas.text_cursor = [0,8]
-canvas.print("0 DEG")
+canvas.text "0 DEG"
 
 # Canvas always rotates around its center, and 0,0 is always the new top-left corner.
 canvas.rotate(90)
 canvas.text_cursor = [0,8]
-canvas.print("90 DEG")
+canvas.text "90 DEG"
 
 # Transforms are cumulative, but do not affect anything previously drawn.
 canvas.rotate(90)
 canvas.text_cursor = [0,8]
-canvas.print("180 DEG")
+canvas.text "180 DEG"
 
 canvas.rotate(90)
 canvas.text_cursor = [0,8]
-canvas.print("270 DEG")
+canvas.text "270 DEG"
 
 # Send the canvas to the OLED's graphics RAM so it shows.
 display.draw