+++ examples/display/hd44780.rb
@@ -32,16 +32,16 @@ heart = [	0b00000,
 # Define the character in CGRAM address 2. 0-7 are usable.
 lcd.create_char(2, heart)
 
-# Need to call home/clear/set_cursor so we go back to writing DDRAM.
+# Need to call home/clear/text_cursor so we go back to writing DDRAM.
 lcd.home
 
 # End the first line with the heart by writing its CGRAM address.
-lcd.print "Hello World!   "
+lcd.text "Hello World!   "
 lcd.write(2)
 
 # Display a clock on second line, updating approximately every second.
 loop do
-  lcd.move_to 0,1
-  lcd.print(Time.now.strftime("%I:%M:%S"))
+  lcd.text_cursor 0,1
+  lcd.text(Time.now.strftime("%I:%M:%S"))
   sleep 1
 end