+++ examples/display/st7302.rb
@@ -16,7 +16,7 @@ baseline = 75
 canvas.text_cursor = [32,baseline+35]
 canvas.font = Denko::Fonts::LED_8x16
 canvas.font_scale = 2
-canvas.print("Hello World!")
+canvas.text "Hello World!"
 
 # Add some shapes.
 canvas.rectangle(20, baseline, 60, -60)